requirements.txt
filelock==3.12.2
h5py==3.9.0
lxml==4.9.1
matplotlib==3.6.3
numpy==1.25.0
pandas==1.4.2
psutil==5.9.5
PySide6-Essentials==6.4.3
python-dateutil==2.8.2
PyYAML==6.0
qasync==0.24.0
tables==3.7.0
vtk==9.2.6
xmlschema==1.10.0